Moved nested /docs to high-level /docs
Retructure repo

Update README.rst
