+++ .dockerignore
@@ -1,0 +1,1 @@
+./ops/nonvoting_testnet
