+++ ops/nonvoting_testnet/conf/auth/authority.toml
@@ -1,0 +1,105 @@
+# Katzenpost non-voting authority configuration file.
+
+#
+# The Authority section contains mandatory information.
+#
+
+[Authority]
+
+  # Addresses are the IP address/port combinations that the authority will bind
+  # to for incoming connections.
+  Addresses = [ "172.28.1.2:29483" ]
+
+  # DataDir is the absolute path to the server's state files.
+  DataDir = "/conf/authority_data"
+
+#
+# The Logging section controls the logging.
+#
+
+[Logging]
+
+  # Disable disables logging entirely.
+  Disable = false
+
+  # File specifies the log file, if omitted stdout will be used.
+  File = "/conf/authority_data/katzenpost.log"
+
+  # Level specifies the log level out of `ERROR`, `WARNING`, `NOTICE`,
+  # `INFO` and `DEBUG`.
+  #
+  # Warning: The `DEBUG` log level is unsafe for production use.
+  Level = "DEBUG"
+
+
+[Debug]
+  MinNodesPerLayer = 1
+
+#
+# The Parameters section holds the network parameters.
+#
+
+[Parameters]
+
+  SendRatePerMinute = 0
+
+  # Mu is the inverse of the mean of the exponential distribution that
+  # clients will use to sample delays.
+  Mu = 0.001
+
+  # MuMaxDelay is the maximum per-hop delay in milliseconds.
+  #
+  # If omitted, the MaxDelay will be derived from the 0.99999 quantile of
+  # of the exponential distribution.
+  MuMaxDelay = 90000
+
+  # LambdaP is the mean of the poisson distribution that clients will use
+  # to sample the send scheduling interval (seconds).
+  LambdaP = 0.0001234
+
+  # LambaPMaxDelay is the maximum send interval in milliseconds.
+  LambdaPMaxDelay = 30000
+
+  # LambdaL is the mean of the poisson distribution that clients will use
+  # to sample the send scheduling interval (seconds).
+  LambdaL = 0.0001234
+
+  # LambaLMaxDelay is the maximum send interval in milliseconds.
+  LambdaLMaxDelay = 30000
+
+  # LambdaD is the mean of the poisson distribution that clients will use
+  # to sample the send scheduling interval (seconds).
+  LambdaD = 0.0001234
+
+  # LambaDMaxDelay is the maximum send interval in milliseconds.
+  LambdaDMaxDelay = 30000
+
+
+#
+# The Mixes array defines the list of white-listed non-provider nodes.
+#
+
+[[Mixes]]
+  # mix1
+  IdentityKey = "4r9jePAbuzhytcnM3nbD5UQOmzl1X1hI8QOMbhTrg8s="
+
+[[Mixes]]
+  # mix2
+  IdentityKey = "1x8tq04hY+i83o9Yn2nrfTkFj4jXIMWCWilR7fgrWyM="
+
+[[Mixes]]
+  # mix3
+  IdentityKey = "La2FnsbKoU8dzQhfN4zkxSYn6T7LRWmFn0JgtlEklQw="
+
+#
+# The Providers array defines the list of white-listed Provider nodes.
+#
+
+[[Providers]]
+  Identifier = "provider2"
+  IdentityKey = "imigzI26tTRXyYLXujLEPI9QrNYOEgC4DElsFdP9acQ="
+
+[[Providers]]
+  Identifier = "provider1"
+  IdentityKey = "2krwfNDfbakZCSTUUZYKXwdduzlEgS9Jfwm7eyZ0sCg="
+
