+++ ops/nonvoting_testnet/conf/auth/authority_data/identity.private.pem
@@ -1,0 +1,4 @@
+-----BEGIN ED25519 PRIVATE KEY-----
+bI4vCmWUlQOupW2Tr/rLbDjzDmE1kL5Qb7doaSpHOWKjjDU3KP+co3CGjlJZ8Ah+
+HtIxTwVHHnacwcaBiwwepA==
+-----END ED25519 PRIVATE KEY-----
