+++ ops/nonvoting_testnet/conf/auth/authority_data/identity.public.pem
@@ -1,0 +1,3 @@
+-----BEGIN ED25519 PUBLIC KEY-----
+o4w1Nyj/nKNwho5SWfAIfh7SMU8FRx52nMHGgYsMHqQ=
+-----END ED25519 PUBLIC KEY-----
