+++ ops/nonvoting_testnet/conf/mix1/mix_data/identity.private.pem
@@ -1,0 +1,4 @@
+-----BEGIN ED25519 PRIVATE KEY-----
+LGH1LpJqJAG+Hg1wV6oMgYtQ838xZS9vRwjQ0B0kybfiv2N48Bu7OHK1yczedsPl
+RA6bOXVfWEjxA4xuFOuDyw==
+-----END ED25519 PRIVATE KEY-----
