+++ ops/nonvoting_testnet/conf/mix1/mix_data/identity.public.pem
@@ -1,0 +1,3 @@
+-----BEGIN ED25519 PUBLIC KEY-----
+4r9jePAbuzhytcnM3nbD5UQOmzl1X1hI8QOMbhTrg8s=
+-----END ED25519 PUBLIC KEY-----
