+++ ops/nonvoting_testnet/conf/mix1/mix_data/link.private.pem
@@ -1,0 +1,3 @@
+-----BEGIN X25519 PRIVATE KEY-----
+Xv4PJ6gPCJEvuLVQ7dsUMT21KnASz4O82CyJhpCPoXg=
+-----END X25519 PRIVATE KEY-----
