+++ ops/nonvoting_testnet/conf/mix2/mix_data/identity.private.pem
@@ -1,0 +1,4 @@
+-----BEGIN ED25519 PRIVATE KEY-----
+ApYndNBRTyY5Peu9sO7fQiZphU04Kepp3ai+dFXDdrLXHy2rTiFj6Lzej1ifaet9
+OQWPiNcgxYJaKVHt+CtbIw==
+-----END ED25519 PRIVATE KEY-----
