+++ ops/nonvoting_testnet/conf/mix2/mix_data/identity.public.pem
@@ -1,0 +1,3 @@
+-----BEGIN ED25519 PUBLIC KEY-----
+1x8tq04hY+i83o9Yn2nrfTkFj4jXIMWCWilR7fgrWyM=
+-----END ED25519 PUBLIC KEY-----
