+++ ops/nonvoting_testnet/conf/mix2/mix_data/link.private.pem
@@ -1,0 +1,3 @@
+-----BEGIN X25519 PRIVATE KEY-----
+MpTalIlC0SqMEWvgqiohvu18vfxxq6Rm5cTVZBFE9eY=
+-----END X25519 PRIVATE KEY-----
