+++ ops/nonvoting_testnet/conf/mix3/mix_data/identity.private.pem
@@ -1,0 +1,4 @@
+-----BEGIN ED25519 PRIVATE KEY-----
+olXV48urFRPjucBUbvZ1i8bDGQozzyaazDndya9OmiAtrYWexsqhTx3NCF83jOTF
+JifpPstFaYWfQmC2USSVDA==
+-----END ED25519 PRIVATE KEY-----
