+++ ops/nonvoting_testnet/conf/mix3/mix_data/identity.public.pem
@@ -1,0 +1,3 @@
+-----BEGIN ED25519 PUBLIC KEY-----
+La2FnsbKoU8dzQhfN4zkxSYn6T7LRWmFn0JgtlEklQw=
+-----END ED25519 PUBLIC KEY-----
