+++ ops/nonvoting_testnet/conf/mix3/mix_data/link.private.pem
@@ -1,0 +1,3 @@
+-----BEGIN X25519 PRIVATE KEY-----
+87rDmtVXd1DW3eqh6Zfs+ICrRu0YT2Rx8CuRs9sIfCE=
+-----END X25519 PRIVATE KEY-----
