+++ ops/nonvoting_testnet/conf/provider1/katzenpost.toml
@@ -1,0 +1,119 @@
+# Katzenpost server configuration file.
+
+#
+# The Server section contains mandatory information common to all nodes.
+#
+
+[Server]
+  Identifier = "provider1"
+
+  # Addresses are the IP address/port combinations that the server will bind
+  # to for incoming connections.
+  Addresses = [ "172.28.1.3:59484"]
+
+  # DataDir is the absolute path to the server's state files.
+  DataDir = "/conf/provider_data"
+
+  # IsProvider specifies if the server is a provider (vs a mix).
+  IsProvider = true
+
+#
+# The PKI section contains the directory authority configuration.
+#
+
+[PKI]
+
+  # Nonvoting is a simple non-voting PKI for mixnets/test deployments.
+  [PKI.Nonvoting]
+
+    # Address is the IP address/port combination of the directory authority.
+    Address = "172.28.1.2:29483"
+
+    # PublicKey is the directory authority's public key in Base64 or Base16 format.
+    PublicKey = "o4w1Nyj/nKNwho5SWfAIfh7SMU8FRx52nMHGgYsMHqQ="
+
+#
+# The Logging section controls the logging.
+#
+
+[Logging]
+
+  # Disable disables logging entirely.
+  Disable = false
+
+  # File specifies the log file, if omitted stdout will be used.
+  File = "/conf/provider_data/katzenpost.log"
+
+  # Level specifies the log level out of `ERROR`, `WARNING`, `NOTICE`,
+  # `INFO` and `DEBUG`.
+  #
+  # Warning: The `DEBUG` log level is unsafe for production use.
+  Level = "DEBUG"
+
+[Debug]
+  DisableRateLimit = true
+
+#
+# The Provider secton specifies the Provider configuration.
+#
+
+[Provider]
+EnableUserRegistrationHTTP = true
+UserRegistrationHTTPAddresses = ["172.28.1.3:36968"]
+AdvertiseUserRegistrationHTTPAddresses = ["http://127.0.0.1:36968"]
+
+[[Provider.Kaetzchen]]
+    Disable = false
+    Capability = "loop"
+    Endpoint = "+loop"
+
+   [[Provider.CBORPluginKaetzchen]]
+     Disable = false
+     Capability = "echo"
+     Endpoint = "+echo"
+     Command = "/go/bin/echo_server"
+     MaxConcurrency = 1
+     [Provider.CBORPluginKaetzchen.Config]
+      log_dir = "/conf/service_logs"
+      log_level = "DEBUG"
+
+   [[Provider.CBORPluginKaetzchen]]
+     Disable = false
+     Capability = "panda"
+     Endpoint = "+panda"
+     Command = "/go/bin/panda_server"
+     MaxConcurrency = 1
+     [Provider.CBORPluginKaetzchen.Config]
+      log_dir = "/conf/service_logs"
+      log_level = "DEBUG"
+      fileStore = "/conf/service_data/panda.storage"
+
+  [[Provider.CBORPluginKaetzchen]]
+    Disable = false
+    Capability = "spool"
+    Endpoint = "+spool"
+    Command = "/go/bin/memspool"
+    MaxConcurrency = 1
+    [Provider.CBORPluginKaetzchen.Config]
+      data_store = "/conf/service_data/memspool.storage"
+      log_dir = "/conf/service_logs"
+
+  [[Provider.CBORPluginKaetzchen]]
+    Disable = false
+    Capability = "eth"
+    Endpoint = "+eth"
+    Command = "/go/bin/Meson"
+    MaxConcurrency = 1
+    [Provider.CBORPluginKaetzchen.Config]
+      log_dir = "/conf/service_logs"
+      log_level = "DEBUG"
+      f = "/conf/currency.toml"
+
+#
+# The Management section specifies the management interface configuration.
+#
+
+[Management]
+
+  # Enable enables the management interface.
+  Enable = false
