+++ ops/nonvoting_testnet/conf/provider1/provider_data/identity.private.pem
@@ -1,0 +1,4 @@
+-----BEGIN ED25519 PRIVATE KEY-----
+ndkH4sYkxVXPFU7OF6wryVar5cWxsZEBcFXWOHnEM3/aSvB80N9tqRkJJNRRlgpf
+B127OUSBL0l/Cbt7JnSwKA==
+-----END ED25519 PRIVATE KEY-----
