+++ ops/nonvoting_testnet/conf/provider1/provider_data/identity.public.pem
@@ -1,0 +1,3 @@
+-----BEGIN ED25519 PUBLIC KEY-----
+2krwfNDfbakZCSTUUZYKXwdduzlEgS9Jfwm7eyZ0sCg=
+-----END ED25519 PUBLIC KEY-----
