+++ ops/nonvoting_testnet/conf/provider1/provider_data/link.private.pem
@@ -1,0 +1,3 @@
+-----BEGIN X25519 PRIVATE KEY-----
+iboFtIVykmdzQoL3rDha5Vs0RtxfmQJT8CyWRbT09Xg=
+-----END X25519 PRIVATE KEY-----
