+++ ops/nonvoting_testnet/conf/provider2/currency.toml
@@ -1,0 +1,7 @@
+Ticker = "ETH"
+RPCUser = "rpcuser"
+RPCPass = "rpcpassword"
+RPCURL = "http://172.28.1.11:9545"
+ChainId = 4
+LogDir = "/conf/service_logs"
+LogLevel = "DEBUG"
