+++ ops/nonvoting_testnet/conf/provider2/katzenpost.toml
@@ -1,0 +1,104 @@
+# Katzenpost server configuration file.
+
+#
+# The Server section contains mandatory information common to all nodes.
+#
+
+[Server]
+  Identifier = "provider2"
+
+  # Addresses are the IP address/port combinations that the server will bind
+  #
+  # to for incoming connections.
+  Addresses = [ "172.28.1.4:29484"]
+
+  # DataDir is the absolute path to the server's state files.
+  DataDir = "/conf/provider_data"
+
+  # IsProvider specifies if the server is a provider (vs a mix).
+  IsProvider = true
+
+#
+# The PKI section contains the directory authority configuration.
+#
+
+[PKI]
+
+  # Nonvoting is a simple non-voting PKI for mixnets/test deployments.
+  [PKI.Nonvoting]
+
+    # Address is the IP address/port combination of the directory authority.
+    Address = "172.28.1.2:29483"
+
+    # PublicKey is the directory authority's public key in Base64 or Base16 format.
+    PublicKey = "o4w1Nyj/nKNwho5SWfAIfh7SMU8FRx52nMHGgYsMHqQ="
+
+#
+# The Logging section controls the logging.
+#
+
+[Logging]
+
+  # Disable disables logging entirely.
+  Disable = false
+
+  # File specifies the log file, if omitted stdout will be used.
+  File = "/conf/provider_data/katzenpost.log"
+
+  # Level specifies the log level out of `ERROR`, `WARNING`, `NOTICE`,
+  # `INFO` and `DEBUG`.
+  #
+  # Warning: The `DEBUG` log level is unsafe for production use.
+  Level = "DEBUG"
+
+[Debug]
+  DisableRateLimit = true
+
+#
+# The Provider secton specifies the Provider configuration.
+#
+
+[Provider]
+EnableUserRegistrationHTTP = true
+UserRegistrationHTTPAddresses = ["172.28.1.4:36967"]
+AdvertiseUserRegistrationHTTPAddresses = ["http://127.0.0.1:36967"]
+
+  [[Provider.Kaetzchen]]
+    Capability = "loop"
+    Endpoint = "+loop"
+    Disable = false
+
+   [[Provider.CBORPluginKaetzchen]]
+     Disable = false
+     Capability = "echo"
+     Endpoint = "+echo"
+     Command = "/go/bin/echo_server"
+     MaxConcurrency = 1
+     [Provider.CBORPluginKaetzchen.Config]
+      log_dir = "/conf/service_logs"
+      log_level = "DEBUG"
+
+  [[Provider.CBORPluginKaetzchen]]
+    Disable = false
+    Capability = "eth"
+    Endpoint = "+eth"
+    Command = "/go/bin/Meson"
+    MaxConcurrency = 1
+    [Provider.CBORPluginKaetzchen.Config]
+      log_dir = "/conf/service_logs"
+      log_level = "DEBUG"
+      f = "/conf/currency.toml"
+
+
+#
+# The Management section specifies the management interface configuration.
+#
+
+[Management]
+
+  # Enable enables the management interface.
+  Enable = true
+
+  # Path specifies the path to the management interface socket.  If left
+  # empty it will use `management_sock` under the DataDir.
+  Path = "/conf/provider_data/management_sock"
