+++ ops/nonvoting_testnet/conf/provider2/provider_data/identity.private.pem
@@ -1,0 +1,4 @@
+-----BEGIN ED25519 PRIVATE KEY-----
+xun4XQfdsh+w8pD+e1Rml9RCaOTfCoZHG3s6OOek9v2KaKDMjbq1NFfJgte6MsQ8
+j1Cs1g4SALgMSWwV0/1pxA==
+-----END ED25519 PRIVATE KEY-----
