+++ ops/nonvoting_testnet/conf/provider2/provider_data/identity.public.pem
@@ -1,0 +1,3 @@
+-----BEGIN ED25519 PUBLIC KEY-----
+imigzI26tTRXyYLXujLEPI9QrNYOEgC4DElsFdP9acQ=
+-----END ED25519 PUBLIC KEY-----
