+++ ops/nonvoting_testnet/conf/provider2/provider_data/link.private.pem
@@ -1,0 +1,3 @@
+-----BEGIN X25519 PRIVATE KEY-----
+04XSX85Ov4jDTb0vqEMv3McYME7weayniGLKtmF6UwQ=
+-----END X25519 PRIVATE KEY-----
